rename to rpc-definition
