+++ apps/game/client/game-rpc-client.ts
@@ -1,7 +1,7 @@
 import type { RouterNode } from "@mp/rpc";
 import type { ReactRpcInvoker } from "@mp/rpc/react";
 import { InjectionContext } from "@mp/ioc";
-import type { gameServerRpcSlice } from "../server/rpc.slice";
+import type { gameServerRpcSlice } from "../server/rpc-definition.slice";
 
 export const ctxGameRpcClient =
   InjectionContext.new<GameRpcClient>("GameRpcClient");
